// dashboard/src/api/models/invoiceDocumentAudienceEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `customer` - Customer
 * `internal` - Internal
 * `legal` - Legal
 */
export type InvoiceDocumentAudienceEnum =
  (typeof InvoiceDocumentAudienceEnum)[keyof typeof InvoiceDocumentAudienceEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceDocumentAudienceEnum = {
  customer: "customer",
  internal: "internal",
  legal: "legal",
} as const;